Reduce "Connecting to Echotest" waiting time by 4 seconds
one of the most annoying things about breakout rooms is that the users have to do the echo test again. So they have to wait 7 to 12 seconds again (for most typical bbb configurations) just for the echo test to connect. This change reduces the iceCheckingTimeout from 5 seconds to 1 second, which is still way more than enough unless there is a configuration or network load issue. Depending on network of the used client, this change saves the user between 4 to 8 seconds for every time they activate audio playback or microphone.
The original 5 seconds timeout is not part of the official sip.js library but instead was added in #2877