Few changes at the reload button
Tweaked a bit at the button's style to give it a little distance from
the borders and replaced by a larger circle button for the icon to appear
a few bigger and keep the roundness of the surrinding design. Also moved
the suppress control to the button component so it solves for both external
player, generic component and any other component that decides to use it in
the future.