fix(audio): add label to device selector button
Some technical information:

Added the possibility to have a tooltip container inside of an emojibutton
(even if the button already has a container). We handle the hover event in the
parent button by adding a simbling element and binding the tooltip in it.
So the tooltip/hover of the ButtonEmoji won't affect the parent element.

No changes, but additions, making the default behavior of tooltip intact,
avoid regressions on it.