fix(breakout rooms): waiting url state
Setting the waiting state false as default was leaving an open gap where
the users still did not have a valid breakout room's join url and it's
join button enabled.

Change the first waiting state to true and handle the join type as the
component finishes mounting. If it's a free to choose room scenario,
dispatch the whole Promise based solution and let it manage the waiting
state as the interval checks for the URL response.